requirements.txt
streamlit==1.27.2
streamlit-extras==0.3.4
pandas==2.1.1
openai==0.28.1
langchain==0.0.306
python-dotenv==1.0.0